Add setting for checking and ignoring AMP files (#35)
* Add setting for checking and ignoring AMP files

If the intercepted file is an AMP file, it can be ignored by WOVN.php and content should not be changed. Because WOVN.php interception is explicit most of the time (user controls what to intercept), I made it an option that is disabled by default.

Because WOVN.php avoids using HTML parser most of the time, and because this feature is disabled by default I am using regexps too.