Change from realpath to our own library for sanitizing input
realpath sanitizes the input by resolving . and .. but
it also returns false, with no path, when the path does
not exist.  But in the case of the code, it was searching
the root directory for an index file, if the root directory
did not exist.  Now it only searches the url given and yields
a 404 error on error.