Support HTTP Header X-Forwarded-Request-Uri for when proxy is used (#89)
When WOVN.php serves a request behind a proxy and the request URI gets modified in the process, it will need to read the original request URI from a HTTP Header. The header name for original URI is `X-Forwarded-Request-Uri`, and it must be manually set by the user. See the README for an example of how set it.

The original client requested URI is needed in order to correctly identify the requested page, and is used for setting hreflang tags in the HTML content and for locating the correct translation data.

WOVN.php will only look at the `X-Forwarded-Request-Uri` header if `use_proxy = 1` is set in wovn.ini.

Version bumped to 0.1.13