Prevent Interception Errors Due to CWD Changed (#96)
* Prevent Interception Errors Due to CWD Changed

When WOVN.php intercepts through wovn_index.php, CWD is the root because
wovn_index.php is at the root. However, the file intercepted by WOVN.php
is not necessary at the root. When intercepted file is not at the root,
import may be relative to the directory of the file. Therefore, when
intercepting files in subdirectories, imports are broken. This change
set the CWD to the directory of the file intercepted.